% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_exec_error.R
\name{capture_exec_warning}
\alias{capture_exec_warning}
\title{Capture an execution warning condition}
\usage{
capture_exec_warning(file_path, msg, call = NULL)
}
\arguments{
\item{file_path}{character string. Path to the file being validated. Must be
the relative path to the hub's \code{model-output} (or equivalent) directory.}

\item{msg}{Character string.}
}
\value{
A \verb{<warning/check_exec_warn>} condition class object. Returned object also
inherits from subclass \verb{<hub_check>}.
}
\description{
Capture an execution warning condition. Useful for communicating when a check
execution has failed. Usually used in conjunction with \code{\link{try}}.
}
