% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_location.R
\name{check_file_location}
\alias{check_file_location}
\title{Check file is being submitted to the correct folder}
\usage{
check_file_location(file_path)
}
\arguments{
\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object
\item \verb{<warning/check_failure>} condition class object
}

Objects also inherit from subclass \verb{<hub_check>}.
}
\description{
Checks that the \code{model_id} metadata in the file name matches the directory name
the file is being submitted to.
}
