% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_errors.R
\name{check_for_errors}
\alias{check_for_errors}
\title{Raise conditions stored in validation objects}
\usage{
check_for_errors(x, verbose = FALSE, show_warnings = FALSE)
}
\arguments{
\item{x}{A \code{hub_validations} or \code{hub_validations_collection} object
(including subclasses \code{target_validations} and
\code{target_validations_collection}).}

\item{verbose}{Logical. If \code{TRUE}, print the results of all checks prior to
raising condition and summarising validation object check results.}

\item{show_warnings}{Logical. If \code{TRUE}, print check-level warnings inline
with their checks. Validation-level warnings are always printed. Default
\code{FALSE}.}
}
\value{
An error if one of the elements of \code{x} is of class \code{check_failure},
\code{check_error}, \code{check_exec_error} or \code{check_exec_warning}.
\code{TRUE} invisibly otherwise.
}
\description{
Checks validation objects for errors and raises conditions if any are found.
Works with \code{hub_validations} and \code{hub_validations_collection} objects, as
well as their subclasses (\code{target_validations} and
\code{target_validations_collection}). Can be used in CI workflows to signal
validation failures, or locally to summarise validation results.
}
\details{
For more details on these classes, see
\href{https://hubverse-org.github.io/hubValidations/articles/hub-validations-class.html}{article on \verb{<hub_validations>} S3 class objects}.
}
\seealso{
\code{\link[=validate_submission]{validate_submission()}}, \code{\link[=validate_pr]{validate_pr()}},
\code{\link[=validate_target_submission]{validate_target_submission()}}, \code{\link[=validate_target_pr]{validate_target_pr()}}
}
