% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata_file_ext.R
\name{check_metadata_file_ext}
\alias{check_metadata_file_ext}
\title{Check file is being submitted to the correct folder}
\usage{
check_metadata_file_ext(file_path)
}
\arguments{
\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Checks that the \code{model_id} metadata in the file name matches the directory name
the file is being submitted to.
}
