% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata_file_location.R
\name{check_metadata_file_location}
\alias{check_metadata_file_location}
\title{Check that the metadata file is being submitted to the correct folder}
\usage{
check_metadata_file_location(file_path)
}
\arguments{
\item{file_path}{character string. Path to the file being validated relative to
the hub's model-metadata directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<warning/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check that the metadata file is being submitted to the correct folder
}
