% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target_dataset_unique.R
\name{check_target_dataset_unique}
\alias{check_target_dataset_unique}
\title{Check that a single unique target dataset exists for a given target type.}
\usage{
check_target_dataset_unique(
  hub_path,
  target_type = c("time-series", "oracle-output")
)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{target_type}{Type of target data to retrieve matching files. One of "time-series" or
"oracle-output". Defaults to "time-series".}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check that a single unique target dataset exists for a given target type.
}
