% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target_file_ext_valid.R
\name{check_target_file_ext_valid}
\alias{check_target_file_ext_valid}
\title{Check that a target data file has a valid extension.}
\usage{
check_target_file_ext_valid(file_path)
}
\arguments{
\item{file_path}{A character string representing the path to the target data file
relative to the \code{target-data} directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Note that files which are part of a hive partitioned dataset must have
parquet file extension only.
}
