% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target_file_name.R
\name{check_target_file_name}
\alias{check_target_file_name}
\title{Check that a hive-partitioned target data file name can be correctly parsed.}
\usage{
check_target_file_name(file_path)
}
\arguments{
\item{file_path}{A character string representing the path to the target data file
relative to the \code{target-data} directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check that a hive-partitioned target data file name can be correctly parsed.
}
