% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target_tbl_ts_targets.R
\name{check_target_tbl_ts_targets}
\alias{check_target_tbl_ts_targets}
\title{Check that targets in a time-series target data file are valid}
\usage{
check_target_tbl_ts_targets(
  target_tbl,
  target_type = c("time-series", "oracle-output"),
  file_path,
  hub_path
)
}
\arguments{
\item{target_tbl}{A tibble/data.frame of the contents of the target data file
being validated.}

\item{target_type}{Type of target data to retrieve matching files. One of "time-series" or
"oracle-output". Defaults to "time-series".}

\item{file_path}{A character string representing the path to the target data file
relative to the \code{target-data} directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check is only performed when the target data type is \code{time-series}.
When the target task ID is not specified in the config (i.e. hub has single
target and \code{target_keys = NULL}), the validity of the target is only checked
through the config file. Otherwise, the values in the target task ID column
of \code{target_tbl} are checked. Note that valid \code{time-series} targets must be
step ahead and their target type must be one of \code{"continuous"}, \code{"discrete"},
\code{"binary"} or \code{"compositional"}. If the hub contains no valid time-series targets, no time-series target data should be present and validation of such data will be skipped.
}
