% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_colnames.R
\name{check_tbl_colnames}
\alias{check_tbl_colnames}
\title{Check column names of model output data}
\usage{
check_tbl_colnames(tbl, round_id, config_tasks, file_path)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{round_id}{character string. The round identifier.}

\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
A list containing one of:
\itemize{
\item \verb{<message/check_success>} condition class object
\item \verb{<warning/check_failure>} condition class object
}

depending on whether validation has succeeded.
}
\description{
Checks that a tibble/data.frame of data read in from the file being validated
contains the expected task ID and standard column names according the round
configuration being validated against.
}
