% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_spl_mt_unique.R
\name{check_tbl_spl_mt_unique}
\alias{check_tbl_spl_mt_unique}
\title{Check that individual sample output_type_ids do not span multiple model tasks}
\usage{
check_tbl_spl_mt_unique(
  tbl_chr,
  round_id,
  file_path,
  hub_path,
  derived_task_ids = get_hub_derived_task_ids(hub_path, round_id)
)
}
\arguments{
\item{tbl_chr}{a tibble/data.frame of the contents of the file being
validated. Column types must \strong{all be character}.}

\item{round_id}{character string. The round identifier.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs) to ignore during validation. Defaults to
extracting derived task IDs from hub \code{task.json}. See
\code{\link[=get_hub_derived_task_ids]{get_hub_derived_task_ids()}} for more details.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check that individual sample output_type_ids do not span multiple model tasks
}
\details{
Different model tasks can have different sample configurations
(\code{compound_taskid_set}, \code{min/max_samples_per_task}, etc.), so samples should
be entirely independent across model tasks. This check verifies that no
sample \code{output_type_id} appears in more than one model task.

Output of the check includes an \code{errors} element, a list with the following
structure:
\itemize{
\item \code{mt_ids}: Integer vector of model task indices the overlapping samples span.
\item \code{output_type_ids}: Character vector of sample \code{output_type_id}s that appear
in multiple model tasks.
}
}
