% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_spl_non_compound_tid.R
\name{check_tbl_spl_non_compound_tid}
\alias{check_tbl_spl_non_compound_tid}
\title{Check model output data tbl samples contain single unique combination of
non-compound task ID values across all samples}
\usage{
check_tbl_spl_non_compound_tid(
  tbl_chr,
  round_id,
  file_path,
  hub_path,
  compound_taskid_set = NULL,
  derived_task_ids = get_hub_derived_task_ids(hub_path, round_id)
)
}
\arguments{
\item{tbl_chr}{a tibble/data.frame of the contents of the file being
validated. Column types must \strong{all be character}.}

\item{round_id}{character string. The round identifier.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{compound_taskid_set}{a list of \code{compound_taskid_set}s (characters vector of compound task IDs),
one for each modeling task. Used to override the compound task ID set in the config file,
for example, when validating coarser samples.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs) to ignore during validation. Defaults to
extracting derived task IDs from hub \code{task.json}. See
\code{\link[=get_hub_derived_task_ids]{get_hub_derived_task_ids()}} for more details.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check model output data tbl samples contain single unique combination of
non-compound task ID values across all samples
}
\details{
Output of the check includes an \code{errors} element, a list of items,
one for each modeling task containing samples failing validation,
with the following structure:
\itemize{
\item \code{mt_id}: Index identifying the config modeling task the samples are associated with.
\item \code{output_type_ids}: The output type IDs of samples that do not match the most frequent
non-compound task ID value combination across all
samples in the modeling task.
\item \code{frequent}: The most frequent non-compound task ID value combination
across all samples in the modeling task to which all samples were compared.
See \href{https://docs.hubverse.io/en/latest/user-guide/sample-output-type.html}{hubverse documentation on samples}
for more details.
}
}
