% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_unique_round_id.R
\name{check_tbl_unique_round_id}
\alias{check_tbl_unique_round_id}
\title{Check model output data tbl contains a single unique round ID.}
\usage{
check_tbl_unique_round_id(tbl, file_path, hub_path, round_id_col = NULL)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{round_id_col}{Character string. The name of the column containing
\code{round_id}s. Usually, the value of round property \code{round_id} in hub \code{tasks.json}
config file.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<warning/check_error>} condition class object.
}

If \code{round_id_from_variable: false} and no \code{round_id_col} name is provided,
check is skipped and a \verb{<message/check_info>} condition class object is
returned. If no valid \code{round_id_col} name is provided or can extracted from
config (check through \code{check_valid_round_id_col}), a \verb{<message/check_warning>}
condition class object is returned and the rest of the check skipped.
}
\description{
Check model output data tbl contains a single unique round ID.
}
\details{
This check only applies to files being submitted to rounds where
\code{round_id_from_variable: true} or where a \code{round_id_col} name is explicitly
provided. Skipped otherwise.
}
