% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_value_col_sum1.R
\name{check_tbl_value_col_sum1}
\alias{check_tbl_value_col_sum1}
\title{Check that \code{pmf} output type values of model output data sum to 1.}
\usage{
check_tbl_value_col_sum1(tbl, file_path)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<warning/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Checks that values in the \code{value} column of \code{pmf} output type
data for each unique task ID combination sum to 1.
Check only performed if \code{tbl} contains \code{pmf} output type data.
If not, the check is skipped and a \verb{<message/check_info>} condition class
object is returned.
}
