% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_values_required.R
\name{check_tbl_values_required}
\alias{check_tbl_values_required}
\title{Check all required task ID/output type/output type ID value combinations present
in model data.}
\usage{
check_tbl_values_required(
  tbl_chr,
  round_id,
  file_path,
  hub_path,
  derived_task_ids = get_hub_derived_task_ids(hub_path, round_id)
)
}
\arguments{
\item{tbl_chr}{a tibble/data.frame of the contents of the file being
validated. Column types must \strong{all be character}.}

\item{round_id}{character string. The round identifier.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs) to ignore during validation. Defaults to
extracting derived task IDs from hub \code{task.json}. See
\code{\link[=get_hub_derived_task_ids]{get_hub_derived_task_ids()}} for more details.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check all required task ID/output type/output type ID value combinations present
in model data.
}
\details{
Note that it is \strong{necessary for \code{derived_task_ids} to be specified if any of
the task IDs with \code{required} values have dependent derived task IDs}. If this is the
case and derived task IDs are not specified, the dependent nature of derived
task ID values will result in \strong{false validation errors when validating
required values}.
}
