% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_round_id.R
\name{check_valid_round_id}
\alias{check_valid_round_id}
\title{Check whether the \code{round_id} determined for the submission is valid}
\usage{
check_valid_round_id(round_id, config_tasks, file_path)
}
\arguments{
\item{round_id}{character string. The round identifier.}

\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
A list containing one of:
\itemize{
\item \verb{<message/check_success>} condition class object
\item \verb{<error/check_error>} condition class object
}

depending on whether validation has succeeded.
}
\description{
Check whether the \code{round_id} determined for the submission is valid
}
