% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_validations_methods.R
\name{combine.hub_validations}
\alias{combine.hub_validations}
\title{Concatenate \code{hub_validations} S3 class objects}
\usage{
\method{combine}{hub_validations}(...)
}
\arguments{
\item{...}{\code{hub_validations} S3 class objects to be concatenated.}
}
\value{
a \code{hub_validations} S3 class object.
}
\description{
Concatenate \code{hub_validations} S3 class objects
}
