% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine.target_validations}
\alias{combine.target_validations}
\title{Concatenate \code{target_validations} S3 class objects}
\usage{
\method{combine}{target_validations}(...)
}
\arguments{
\item{...}{\code{target_validations} S3 class objects to be concatenated.}
}
\value{
a \code{target_validations} S3 class object.
}
\description{
Concatenate \code{target_validations} S3 class objects
}
