% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_check.R
\name{create_custom_check}
\alias{create_custom_check}
\title{Create a custom validation check function template file.}
\usage{
create_custom_check(
  name,
  hub_path = ".",
  r_dir = "src/validations/R",
  error = FALSE,
  conditional = FALSE,
  error_object = FALSE,
  extra_args = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{name}{Character string. Name of the custom check function. We recommend
following the hubValidations package naming convention. For more details, consult the
article on \href{https://hubverse-org.github.io/hubValidations/articles/writing-custom-fns.html}{writing custom check functions}.}

\item{hub_path}{Character string. Path to the hub directory. Default is the
current working directory.}

\item{r_dir}{Character string. Path (relative to \code{hub_path}) to the directory
the custom check function file will be written to. Default is \code{src/validations/R}
which is the recommended directory for storing custom check functions.}

\item{error}{Logical. If \code{TRUE}, the custom check function will return a
\verb{<error/check_error>} class object in the case of a failed check instead of the
default \verb{<error/check_failure>}.}

\item{conditional}{Logical. If \code{TRUE}, the custom check function template will
include a block of code to check a condition before running the check. This is useful
when a check may need to be skipped based on a condition.}

\item{error_object}{Logical. If \code{TRUE}, the custom check function template will
include an error object that can be used to store additional information about the
properties of the object being checked that caused check failure. For example,
it could store the index of rows in a \code{tbl} that caused a check failure.}

\item{extra_args}{Logical. If \code{TRUE}, the custom check function template will
include an \code{extra_arg} template argument and template block of code to check
the input arguments of the custom check function.}

\item{overwrite}{Logical. If \code{TRUE}, the function will overwrite an existing}
}
\value{
Invisible \code{TRUE} if the custom check function file is created successfully.
}
\description{
Create a custom validation check function template file.
}
\examples{
withr::with_tempdir({
  # Create the custom check file with default settings.
  create_custom_check("check_default")
  cat(readLines("src/validations/R/check_default.R"), sep = "\n")

 # Create fully featured custom check file.
  create_custom_check("check_full",
    error = TRUE, conditional = TRUE,
    error_object = TRUE, extra_args = TRUE
  )
  cat(readLines("src/validations/R/check_full.R"), sep = "\n")
})
}
