% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_config.R
\name{get_config_derived_task_ids}
\alias{get_config_derived_task_ids}
\title{Get hub configuration fields from a \verb{<config>} class object}
\usage{
get_config_derived_task_ids(config_tasks, round_id = NULL)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[hubUtils:read_config]{hubUtils::read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the round's
\code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config. Ignored if hub
contains only a single round.}
}
\value{
\itemize{
\item \code{get_config_derived_task_ids}: character vector of hub or round level derived
task ID names. If \code{round_id} is \code{NULL} or the round does not have a round level
\code{derived_tasks_ids} setting, returns the hub level \code{derived_tasks_ids} setting.
}
}
\description{
Get hub configuration fields from a \verb{<config>} class object
}
\section{Functions}{
\itemize{
\item \code{get_config_derived_task_ids()}: Get the hub or round level \code{derived_tasks_ids}

}}
\examples{
hub_path <- system.file("testhubs/v4/flusight", package = "hubUtils")
config_tasks <- read_config(hub_path)
get_config_derived_task_ids(config_tasks)
get_config_derived_task_ids(config_tasks, round_id = "2023-05-08")
}
