% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{get_filenames}
\alias{get_filenames}
\title{Get filenames from hub validation object and it's subclasses}
\usage{
get_filenames(x, unique = FALSE, ...)
}
\arguments{
\item{x}{A validation object or one of it's subclasses}

\item{unique}{Logical, whether to return unique filenames only}

\item{...}{Additional arguments passed to methods}
}
\description{
Get filenames from hub validation object and it's subclasses
}
