% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-target-metadata.R
\name{get_target_task_id}
\alias{get_target_task_id}
\title{Get Unique Target Task ID}
\usage{
get_target_task_id(config_tasks)
}
\arguments{
\item{config_tasks}{a list representation of the \code{tasks.json} config file.}
}
\value{
A character vector of unique target task ID names. Post v5.0.0
this should be a single task ID name.
}
\description{
Retrieves the unique target task ID by extracting all target metadata
and extracting the names of their \code{target_keys}. For valid config files
these should be the same across all rounds and model tasks.
}
