% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_taskid-utils.R
\name{get_tbl_compound_taskid_set}
\alias{get_tbl_compound_taskid_set}
\title{Detect the compound_taskid_set for a tbl for each modeling task in a given round.}
\usage{
get_tbl_compound_taskid_set(
  tbl_chr,
  config_tasks,
  round_id,
  compact = TRUE,
  error = TRUE,
  derived_task_ids = get_config_derived_task_ids(config_tasks, round_id)
)
}
\arguments{
\item{tbl_chr}{a tibble/data.frame of the contents of the file being
validated. Column types must \strong{all be character}.}

\item{config_tasks}{a list representantion of the \code{tasks.json} config file.}

\item{round_id}{Character string. The round ID.}

\item{compact}{Logical. If TRUE, the output will be compacted to remove NULL elements.}

\item{error}{Logical. If TRUE, an error will be thrown if the compound task ID set is not valid.
If FALSE and an error is detected, the detected compound task ID set will be
returned with error attributes attached.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs) to ignore. Columns for such task ids will
contain \code{NA}s. Defaults to extracting derived task IDs from \code{config_tasks}. See
\code{\link[=get_config_derived_task_ids]{get_config_derived_task_ids()}} for more details.}
}
\value{
A list of vectors of compound task IDs detected in the tbl, one for each
modeling task in the round. If \code{compact} is TRUE, modeling tasks returning NULL
elements will be removed.
}
\description{
Detect the compound_taskid_set for a tbl for each modeling task in a given round.
}
\examples{
hub_path <- system.file("testhubs/samples", package = "hubValidations")
file_path <- "flu-base/2022-10-22-flu-base.csv"
round_id <- "2022-10-22"
tbl_chr <- read_model_out_file(
  file_path = file_path,
  hub_path = hub_path,
  coerce_types = "chr"
)
config_tasks <- read_config(hub_path, "tasks")
get_tbl_compound_taskid_set(tbl_chr, config_tasks, round_id)
get_tbl_compound_taskid_set(tbl_chr, config_tasks, round_id,
  compact = FALSE
)

}
