% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubValidations-package.R
\docType{package}
\name{HubValidations-package}
\alias{HubValidations}
\alias{HubValidations-package}
\title{HubValidations: Validate Data Submitted to a Modeling Hub}
\description{
This package aims at providing a simple interface to run validations on data and metadata submitted to a modeling hub. Validation tests can be run at different levels (single file, single folder, whole repository) and locally as well as part of a continuous integration workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiforecasts/HubValidations}
  \item Report bugs at \url{https://github.com/epiforecasts/HubValidations/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]

}
\keyword{internal}
