% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_utils.R
\name{is_success}
\alias{is_success}
\alias{is_failure}
\alias{is_error}
\alias{is_info}
\alias{not_pass}
\title{Get status of a hub check}
\usage{
is_success(x)

is_failure(x)

is_error(x)

is_info(x)

not_pass(x)
}
\arguments{
\item{x}{an object that inherits from class \verb{<hub_check>} to test.}
}
\value{
Logical. Is given status of check TRUE?
}
\description{
Get status of a hub check
}
\section{Functions}{
\itemize{
\item \code{is_success()}: Is check success?

\item \code{is_failure()}: Is check failure?

\item \code{is_error()}: Is check error?

\item \code{is_info()}: Is check info?

\item \code{not_pass()}: Did check not pass?

}}
