% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_hub_validations.R
\name{new_hub_validations}
\alias{new_hub_validations}
\alias{as_hub_validations}
\title{Create new or convert list to \code{hub_validations} S3 class object}
\usage{
new_hub_validations(...)

as_hub_validations(x)
}
\arguments{
\item{...}{named elements to be included. Each element must be an object which
inherits from class \verb{<hub_check>}.}

\item{x}{a list of named elements. Each element must be an object which
inherits from class \verb{<hub_check>}.}
}
\value{
an S3 object of class \verb{<hub_validations>}.
}
\description{
Create new or convert list to \code{hub_validations} S3 class object
}
\section{Functions}{
\itemize{
\item \code{new_hub_validations()}: Create new \verb{<hub_validations>} S3 class object

\item \code{as_hub_validations()}: Convert list to \verb{<hub_validations>} S3 class object

}}
\examples{
new_hub_validations()

hub_path <- system.file("testhubs/simple", package = "hubValidations")
file_path <- "team1-goodmodel/2022-10-08-team1-goodmodel.csv"
new_hub_validations(
  file_exists = check_file_exists(file_path, hub_path),
  file_name = check_file_name(file_path)
)
x <- list(
  file_exists = check_file_exists(file_path, hub_path),
  file_name = check_file_name(file_path)
)
as_hub_validations(x)
}
