% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_target_validations.R
\name{new_target_validations}
\alias{new_target_validations}
\alias{as_target_validations}
\title{Create new or convert list to \code{target_validations} S3 class object}
\usage{
new_target_validations(...)

as_target_validations(x)
}
\arguments{
\item{...}{named elements to be included. Each element must be an object which
inherits from class \verb{<hub_check>}. All checks must have the same \verb{$where} value.}

\item{x}{a list of named elements. Each element must be an object which
inherits from class \verb{<hub_check>}. All checks must have the same \verb{$where} value.}
}
\value{
an S3 object of class \verb{<target_validations>} with a \code{where} attribute.
}
\description{
A \code{target_validations} object contains validation results for a single
validation subject. Depending on context, this could be a target data file,
the hub configuration directory (\code{hub-config}), or a target dataset type
(\code{time-series}, \code{oracle-output}).
All checks must have the same \verb{$where} value, which is extracted and stored
as the \code{where} attribute.
}
\section{Functions}{
\itemize{
\item \code{new_target_validations()}: Create new \verb{<target_validations>} S3 class object

\item \code{as_target_validations()}: Convert list to \verb{<target_validations>} S3 class object

}}
\examples{
new_target_validations()

hub_path <- system.file("testhubs/v5/target_file", package = "hubUtils")
file_path <- "time-series.csv"
new_target_validations(
  target_file_name = check_target_file_name(file_path),
  target_file_ext_valid = check_target_file_ext_valid(file_path)
)
x <- list(
  target_file_name = check_target_file_name(file_path),
  target_file_ext_valid = check_target_file_ext_valid(file_path)
)
as_target_validations(x)
file_path <- "time-series/target=flu_hosp_rate/part-0.parquet"
new_target_validations(
  target_file_name = check_target_file_name(file_path),
  target_file_ext_valid = check_target_file_ext_valid(file_path)
)
}
