% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_check_metadata_team_max_model_n.R
\name{opt_check_metadata_team_max_model_n}
\alias{opt_check_metadata_team_max_model_n}
\title{Check that submitting team does not exceed maximum number of allowed models
per team}
\usage{
opt_check_metadata_team_max_model_n(file_path, hub_path, n_max = 2L)
}
\arguments{
\item{file_path}{character string. Path to the file being validated relative to
the hub's model-metadata directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{n_max}{Integer. Number of maximum allowed models per team.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<warning/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check that submitting team does not exceed maximum number of allowed models
per team
}
\details{
Should be deployed as part of \code{validate_model_metadata} optional checks.
}
