% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_check_tbl_counts_lt_popn.R
\name{opt_check_tbl_counts_lt_popn}
\alias{opt_check_tbl_counts_lt_popn}
\title{Check that predicted values per location are less than total location population.}
\usage{
opt_check_tbl_counts_lt_popn(
  tbl,
  file_path,
  hub_path,
  targets = NULL,
  popn_file_path = "auxiliary-data/locations.csv",
  popn_col = "population",
  location_col = "location"
)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{targets}{Either a single target key list or a list of multiple target key lists.}

\item{popn_file_path}{Character string.
Path to population data relative to the hub root.
Defaults to \code{auxiliary-data/locations.csv}.}

\item{popn_col}{Character string.
The name of the population size column in the population data set.}

\item{location_col}{Character string.
The name of the location column.
Used to join population data to submission file data.
Must be shared by both files.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check that predicted values per location are less than total location population.
}
\details{
Should only be applied to rows containing count predictions. Use argument
\code{targets} to filter \code{tbl} data to appropriate count target rows.

Should be deployed as part of \code{validate_model_data} optional checks.
}
\examples{
hub_path <- system.file("testhubs/flusight", package = "hubValidations")
file_path <- "hub-ensemble/2023-05-08-hub-ensemble.parquet"
tbl <- hubValidations::read_model_out_file(file_path, hub_path)
# Single target key list
targets <- list("target" = "wk ahead inc flu hosp")
opt_check_tbl_counts_lt_popn(tbl, file_path, hub_path, targets = targets)
}
