% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhub_validations_methods.R
\name{print.fhub_validations}
\alias{print.fhub_validations}
\title{Print results of \code{validate_...()} function as a bullet list}
\usage{
\method{print}{fhub_validations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fhub_validations}}

\item{...}{Unused argument present for class consistency}
}
\description{
Print results of \code{validate_...()} function as a bullet list
}
