% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pr_hub_validations}
\alias{print.pr_hub_validations}
\title{Print results of \code{validate_pr()} function as a bullet list}
\usage{
\method{print}{pr_hub_validations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pr_hub_validations}}

\item{...}{Unused argument present for class consistency}
}
\description{
Print results of \code{validate_pr()} function as a bullet list
}
\keyword{internal}
