% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_out_file.R
\name{read_model_out_file}
\alias{read_model_out_file}
\title{Read a model output file}
\usage{
read_model_out_file(file_path, hub_path = ".", use_hub_schema = TRUE)
}
\arguments{
\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{use_hub_schema}{Logical. When reading in \code{csv} files, whether to use
the hub's schema to specify column data types.}
}
\value{
a tibble of contents of the model output file.
}
\description{
Read a model output file
}
