% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_target_file.R
\name{read_target_file}
\alias{read_target_file}
\title{Read a single target data file}
\usage{
read_target_file(
  target_file_path,
  hub_path,
  coerce_types = c("hub", "chr", "none"),
  date_col = NULL,
  na = c("NA", "")
)
}
\arguments{
\item{target_file_path}{Character string. Path to the target data file being validated
relative to the hub's \code{target-data} directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{coerce_types}{character string. What to coerce column types to on read.
\itemize{
\item \code{hub}: (default) read in (\code{csv}) or coerce (\code{parquet}) to schema according to
\code{hub} config (See \code{\link[hubData:create_timeseries_schema]{hubData::create_timeseries_schema()}} and
\code{\link[hubData:create_oracle_output_schema]{hubData::create_oracle_output_schema()}} for details).
When coercing data types using the \code{hub} schema, the \code{output_type_id_datatype}
can also be used to set the \code{output_type_id} column data type manually.
\item \code{chr}: read in (\code{csv}) or coerce (\code{parquet}) all columns to character.
\item \code{none}: No coercion. Use \code{arrow} \verb{read_*} function defaults.
}}

\item{date_col}{Optional column name to be interpreted as date. Default is \code{NULL}.
Useful when the required date column is a partitioning column in the target data
and does not have the same name as a date typed task ID variable in the config.}

\item{na}{A character vector of strings to interpret as missing values. Only
applies to CSV files. The default is \code{c("NA", "")}. Useful when actual character
string \code{"NA"} values are used in the data. In such a case, use empty cells to
indicate missing values in your files and set \code{na = ""}.}
}
\value{
a tibble of contents of the target data file.
}
\description{
Read a single target data file
}
\examples{
# download example hub
hub_path <- withr::local_tempdir()
example_hub <- "https://github.com/hubverse-org/example-complex-forecast-hub.git"
git2r::clone(url = example_hub, local_path = hub_path)
# read in time-series file
read_target_file("time-series.csv", hub_path)
read_target_file("time-series.csv", hub_path, coerce_types = "chr")
# read in oracle-output file
read_target_file("oracle-output.csv", hub_path)
read_target_file("oracle-output.csv", hub_path, coerce_types = "chr")
}
