% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model.R
\name{validate_model}
\alias{validate_model}
\title{Validate data and metadata for a given model/folder}
\usage{
validate_model(
  model_name,
  data_folder = "data-processed",
  metadata_folder = "model-metadata",
  data_schema = "schema-data.yml",
  metadata_schema = "schema-metadata.yml"
)
}
\arguments{
\item{model_name}{Path to the folder containing the model data}

\item{data_folder}{The path to the folder containing forecasts}

\item{metadata_folder}{The path to the folder containing metadata}

\item{data_schema}{Path to the \code{.yml} schema file}

\item{metadata_schema}{Path to the \code{.yml} schema file}
}
\description{
Validate data and metadata for a given model/folder
}
\examples{
validate_model(
  "example-model",
  system.file("testdata", "data-processed",
              package = "HubValidations"),
  system.file("testdata", "model-metadata",
              package = "HubValidations"),
  system.file("testdata", "schema-data.yml",
              package = "HubValidations"),
  system.file("testdata", "schema-metadata.yml",
              package = "HubValidations")
)
}
