% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file.R
\name{validate_model_file}
\alias{validate_model_file}
\title{Valid properties of a file.}
\usage{
validate_model_file(hub_path, file_path)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
An object of class \code{hub_validations}. Each named element contains
a \code{hub_check} class object reflecting the result of a given check. Function
will return early if a check returns an error.
}
\description{
Valid properties of a file.
}
\examples{
hub_path <- system.file("testhubs/simple", package = "hubUtils")
validate_model_file(hub_path,
  file_path = "team1-goodmodel/2022-10-08-team1-goodmodel.csv"
)
validate_model_file(hub_path,
  file_path = "team1-goodmodel/2022-10-15-team1-goodmodel.csv"
)
}
