% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_forecast.R
\name{validate_model_forecast}
\alias{validate_model_forecast}
\title{Validate model forecast file}
\usage{
validate_model_forecast(forecast_file, forecast_schema, ...)
}
\arguments{
\item{forecast_file}{Path to the forecast \code{.csv} file}

\item{forecast_schema}{Path to the \code{.yml} schema file}
}
\description{
Validate model forecast file
}
\examples{
validate_model_forecast(
  system.file("testdata", "example-model", "2021-07-26-example-model.csv",
              package = "ForecastHubValidations"),
  system.file("testdata", "forecast-schema.yml",
              package = "ForecastHubValidations")
)
}
