% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_metadata.R
\name{validate_model_metadata}
\alias{validate_model_metadata}
\title{Validate model metadata file}
\usage{
validate_model_metadata(metadata_file, schema_file)
}
\arguments{
\item{metadata_file}{Path to the metadata \code{.yml} file}

\item{schema_file}{Path to the \code{.yml} schema file}
}
\description{
Validate model metadata file
}
\details{
Checks:
\itemize{
\item \code{metadata_file} uses \code{.yml} extension
\item \code{metadata_file} name starts with the content of \code{model_abbr} field
\item \code{metadata_file} matches specification from \code{schema_file}
}
}
\examples{
validate_model_metadata(
  system.file("extdata", "metadata-example-model.yml",
              package = "ForecastHubValidations"),
  system.file("extdata", "metadata-schema.yml",
              package = "ForecastHubValidations")
)

}
