% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pr.R
\name{validate_pr}
\alias{validate_pr}
\title{Validate Pull Request}
\usage{
validate_pr(
  hub_path = ".",
  gh_repo,
  pr_number,
  round_id_col = NULL,
  output_type_id_datatype = c("from_config", "auto", "character", "double", "integer",
    "logical", "Date"),
  validations_cfg_path = NULL,
  skip_submit_window_check = FALSE,
  file_modification_check = c("error", "failure", "warn", "message", "none"),
  allow_submit_window_mods = TRUE,
  submit_window_ref_date_from = c("file", "file_path"),
  derived_task_ids = NULL
)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{gh_repo}{GitHub repository address in the format \code{username/repo}}

\item{pr_number}{Number of the pull request to validate}

\item{round_id_col}{Character string. The name of the column containing
\code{round_id}s. Only required if files contain a column that contains \code{round_id}
details but has not been configured via \code{round_id_from_variable: true} and
\verb{round_id:} in in hub \code{tasks.json} config file.}

\item{output_type_id_datatype}{character string. One of \code{"from_config"}, \code{"auto"},
\code{"character"}, \code{"double"}, \code{"integer"}, \code{"logical"}, \code{"Date"}.
Defaults to \code{"from_config"} which uses the setting in the \code{output_type_id_datatype}
property in the \code{tasks.json} config file if available. If the property is
not set in the config, the argument falls back to \code{"auto"} which determines
the  \code{output_type_id} data type automatically from the \code{tasks.json}
config file as the simplest data type required to represent all output
type ID values across all output types in the hub.
When only point estimate output types (where \code{output_type_id}s are \code{NA},) are
being collected by a hub, the \code{output_type_id} column is assigned a \code{character}
data type when auto-determined.
Other data type values can be used to override automatic determination.
Note that attempting to coerce \code{output_type_id} to a data type that is
not valid for the data (e.g. trying to coerce\code{"character"} values to
\code{"double"}) will likely result in an error or potentially unexpected
behaviour so use with care.}

\item{validations_cfg_path}{Path to \code{validations.yml} file. If \code{NULL}
defaults to \code{hub-config/validations.yml}.}

\item{skip_submit_window_check}{Logical. Whether to skip the submission window check.}

\item{file_modification_check}{Character string. Whether to perform check and what to
return when modification/deletion of a previously submitted model output file
or deletion of a previously submitted model metadata file is detected in PR:
\itemize{
\item \code{"error"}: Appends a \verb{<error/check_error>} condition class object for each
applicable modified/deleted file.
\item \code{"warning"}: Appends a \verb{<error/check_failure>} condition class object for each
applicable modified/deleted file.
\item \code{"message"}: Appends a \verb{<message/check_info>} condition class object for each
applicable modified/deleted file.
\item \code{"none"}: No modification/deletion checks performed.
}}

\item{allow_submit_window_mods}{Logical. Whether to allow modifications/deletions
of model output files within their submission windows. Defaults to \code{TRUE}.}

\item{submit_window_ref_date_from}{whether to get the reference date around
which relative submission windows will be determined from the file's
\code{file_path} round ID or the \code{file} contents themselves.
\code{file} requires that the file can be read.
Only applicable when a round is configured to determine the submission
windows relative to the value in a date column in model output files.
Not applicable when explicit submission window start and end dates are
provided in the hub's config.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs). The required values and sample checks
leave these columns out.
If \code{NULL}, defaults to extracting derived task IDs from hub \code{task.json}. See
\code{\link[=get_hub_derived_task_ids]{get_hub_derived_task_ids()}} for more details.}
}
\value{
An object of class \code{hub_validations_collection}, a collection of
validation results. The collection includes entries for hub config
validation (\code{"hub-config"}) and file-specific validations (named by
file path).
}
\description{
Validates model output and model metadata files in a Pull Request.
}
\details{
Only model output and model metadata files are individually validated using
\code{validate_submission()} or \code{validate_model_metadata()} respectively although
as part of checks, hub config files are also validated.
Any other files included in the PR are ignored but flagged in a message.

By default, modifications (which include renaming) and deletions of
previously submitted model output files and deletions or renaming of
previously submitted model metadata files are not allowed
and return a \verb{<error/check_error>} condition class object for each
applicable modified/deleted file. This behaviour can be modified through
arguments \code{file_modification_check}, which controls whether modification/deletion
checks are performed and what is returned if modifications/deletions are detected,
and \code{allow_submit_window_mods}, which controls whether modifications/deletions
of model output files are allowed within their submission windows.

When modification/deletion checks are enabled, each affected file creates an
entry in the returned collection named by the file's path. The check within
each entry is named \code{valid_file_status} (reflecting that we validate the
file's git status).
For example, to access the check for a deleted file:
\code{collection[["team1-goodmodel/2022-10-15-team1-goodmodel.csv"]][["valid_file_status"]]}.

Note that to establish \strong{relative} submission windows when performing
modification/deletion checks and \code{allow_submit_window_mods}
is \code{TRUE}, the reference date is taken as the \code{round_id} extracted from
the file path (i.e. \code{submit_window_ref_date_from} is always set to \code{"file_path"}).
This is because we cannot extract dates from columns of deleted
files. If hub submission window reference dates do not match round IDs in file paths,
currently \code{allow_submit_window_mods} will not work correctly and is best set
to \code{FALSE}. This only relates to hubs/rounds where submission windows are
determined relative to a reference date and not when explicit submission
window start and end dates are provided in the config.

Finally, note that it is \strong{necessary for \code{derived_task_ids} to be specified if any
task IDs with \code{required} values have dependent derived task IDs}. If this
is the case and derived task IDs are not specified, the dependent nature of
derived task ID values will result in \strong{false validation errors when
validating required values}.
\subsection{Checks on model output files}{

Details of checks performed by \code{validate_submission()}\if{html}{\out{<table class="table table-striped table-hover table-condensed table-responsive" style="margin-left: auto; margin-right: auto;">}}\if{html}{\out{
 <thead>
  <tr>
   <th style="text-align:left;"> Name </th>
   <th style="text-align:left;"> Check </th>
   <th style="text-align:left;"> Early return </th>
   <th style="text-align:left;"> Fail output </th>
   <th style="text-align:left;"> Extra info </th>
  </tr>
 </thead>
<tbody>
  <tr>
   <td style="text-align:left;font-weight: bold;"> valid_config </td>
   <td style="text-align:left;"> Hub config valid </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> submission_time </td>
   <td style="text-align:left;"> Current time within file submission window </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> file_exists </td>
   <td style="text-align:left;"> File exists at `file_path` provided </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> file_name </td>
   <td style="text-align:left;"> File name valid </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> file_location </td>
   <td style="text-align:left;"> File located in correct team directory </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> round_id_valid </td>
   <td style="text-align:left;"> File round ID is valid hub round IDs </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> file_format </td>
   <td style="text-align:left;"> File format is accepted hub/round format </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> file_n </td>
   <td style="text-align:left;"> Number of submission files per round per team does not exceed allowed number </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> metadata_exists </td>
   <td style="text-align:left;"> Model metadata file exists in expected location </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> file_read </td>
   <td style="text-align:left;"> File can be read without errors </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> valid_round_id_col </td>
   <td style="text-align:left;"> Round ID var from config exists in data column names. Skipped if `round_id_from_var` is FALSE in config. </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> unique_round_id </td>
   <td style="text-align:left;"> Round ID column contains a single unique round ID. Skipped if `round_id_from_var` is FALSE in config. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> match_round_id </td>
   <td style="text-align:left;"> Round ID from file contents matches round ID from file name. Skipped if `round_id_from_var` is FALSE in config. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> colnames </td>
   <td style="text-align:left;"> File column names match expected column names for round (i.e. task ID names + hub standard column names) </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> col_types </td>
   <td style="text-align:left;"> File column types match expected column types from config. Mainly applicable to parquet &amp; arrow files. </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> valid_vals </td>
   <td style="text-align:left;"> Columns (excluding the `value` column) contain valid combinations of task ID / output type / output type ID values. Derived task ID values are not checked against the task IDs they are derived from. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;"> error_tbl: table of invalid task ID/output type/output type ID value combinations </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> rows_unique </td>
   <td style="text-align:left;"> Columns (excluding the `value` and any derived task ID columns) contain unique combinations of task ID / output type / output type ID values </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> req_vals </td>
   <td style="text-align:left;"> Columns (excluding the `value` and any derived task ID columns) contain all required combinations of task ID / output type / output type ID values </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;"> missing_df: table of missing task ID/output type/output type ID value combinations </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> value_col_valid </td>
   <td style="text-align:left;"> Values in `value` column are coercible to data type configured for each output type </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> value_col_non_desc </td>
   <td style="text-align:left;"> Values in `value` column are non-decreasing as output_type_ids increase for all unique task ID /output type value combinations. Applies to `quantile` or `cdf` output types only </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;"> error_tbl: table of rows affected </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> value_col_sum1 </td>
   <td style="text-align:left;"> Values in the `value` column of `pmf` output type data for each unique task ID combination sum to 1. </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;"> error_tbl: table of rows affected </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> spl_mt_unique </td>
   <td style="text-align:left;"> Individual sample output_type_ids do not span multiple model tasks. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;"> errors: list with mt_ids (model task indices the overlapping samples span) and output_type_ids (sample IDs appearing in multiple model tasks). </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> spl_compound_taskid_set </td>
   <td style="text-align:left;"> Sample compound task id sets for each modeling task match or are coarser than the expected set defined in tasks.json config. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;"> errors: list containing item for each failing modeling task. Exact structure dependent on type of validation failure. See check function documentation for more details. </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> spl_compound_tid </td>
   <td style="text-align:left;"> Samples contain single unique values for each compound task ID within individual samples (v3 and above schema only). </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;"> errors: list containing item for each sample failing validation with breakdown of unique values for each compound task ID. </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> spl_non_compound_tid </td>
   <td style="text-align:left;"> Samples contain single unique combination of non-compound task ID values across all samples (v3 and above schema only). </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;"> errors: list containing item for each modeling task with vectors of output type ids of samples failing validation and example table of most frequent non-compound task ID value combination
across all samples in the modeling task </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> spl_n </td>
   <td style="text-align:left;"> Number of samples for a given compound idx falls within accepted compound task range (v3 and above schema only). </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;"> errors: list containing item for each compound_idx failing validation with sample count, metadata on expected samples and example table of expected structure for samples belonging to
the compound idx in question. </td>
  </tr>
</tbody>
</table>
}}

}

\subsection{Checks on model metadata files}{

Details of checks performed by \code{validate_model_metadata()}\if{html}{\out{<table class="table table-striped table-hover table-condensed table-responsive" style="margin-left: auto; margin-right: auto;">}}\if{html}{\out{
 <thead>
  <tr>
   <th style="text-align:left;"> Name </th>
   <th style="text-align:left;"> Check </th>
   <th style="text-align:left;"> Early return </th>
   <th style="text-align:left;"> Fail output </th>
   <th style="text-align:left;"> Extra info </th>
   <th style="text-align:left;"> optional </th>
  </tr>
 </thead>
<tbody>
  <tr>
   <td style="text-align:left;font-weight: bold;"> metadata_schema_exists </td>
   <td style="text-align:left;"> A model metadata schema file exists in `hub-config` directory. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> metadata_file_exists </td>
   <td style="text-align:left;"> A file with name provided to argument `file_path` exists at the expected location (the `model-metadata` directory). </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> metadata_file_ext </td>
   <td style="text-align:left;"> The metadata file has correct extension (yaml or yml). </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> metadata_file_location </td>
   <td style="text-align:left;"> The metadata file has been saved to correct location. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> metadata_matches_schema </td>
   <td style="text-align:left;"> The contents of the metadata file match the hub's model metadata schema </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> metadata_file_name </td>
   <td style="text-align:left;"> The metadata filename matches the model ID specified in the contents of the file. </td>
   <td style="text-align:left;"> TRUE </td>
   <td style="text-align:left;"> check_error </td>
   <td style="text-align:left;">  </td>
   <td style="text-align:left;"> FALSE </td>
  </tr>
  <tr>
   <td style="text-align:left;font-weight: bold;"> NA </td>
   <td style="text-align:left;"> The number of metadata files submitted by a single team does not exceed the maximum number allowed. </td>
   <td style="text-align:left;"> FALSE </td>
   <td style="text-align:left;"> check_failure </td>
   <td style="text-align:left;">  </td>
   <td style="text-align:left;"> TRUE </td>
  </tr>
</tbody>
</table>
}}

}
}
\examples{
\dontrun{
validate_pr(
  hub_path = ".",
  gh_repo = "hubverse-org/ci-testhub-simple",
  pr_number = 3
)
}
}
