% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_repository.R
\name{validate_repository}
\alias{validate_repository}
\title{Validate a complete repository containing multiple forecast folders}
\usage{
validate_repository(
  data_folder = "data-processed",
  metadata_folder = "model-metadata",
  data_schema = file.path(data_folder, "schema-data.yml"),
  metadata_schema = file.path(data_folder, "schema-metadata.yml")
)
}
\arguments{
\item{data_folder}{The path to the folder containing forecasts}

\item{metadata_folder}{The path to the folder containing metadata}

\item{data_schema}{Path to the \code{.yml} schema file}

\item{metadata_schema}{Path to the \code{.yml} schema file}
}
\value{
An object of class \code{fhub_validations}.
}
\description{
Validate a complete repository containing multiple forecast folders
}
\examples{
validate_repository(
  system.file("testdata", package = "ForecastHubValidations")
)

}
