% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_submission_time.R
\name{validate_submission_time}
\alias{validate_submission_time}
\title{Validate a submitted model data file submission time.}
\usage{
validate_submission_time(
  hub_path,
  file_path,
  ref_date_from = c("file_path", "file")
)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{ref_date_from}{whether to get the reference date around
which relative submission windows will be determined from the file's
\code{file_path} round ID or the \code{file} contents themselves.
\code{file} requires that the file can be read.
Only applicable when a round is configured to determine the submission
windows relative to the value in a date column in model output files.
Not applicable when explicit submission window start and end dates are
provided in the hub's config.}
}
\value{
An object of class \code{hub_validations}. Each named element contains
a \code{hub_check} class object reflecting the result of a given check. Function
will return early if a check returns an error.

For more details on the structure of \verb{<hub_validations>} objects, including
how to access more information on individual checks,
see \href{https://infectious-disease-modeling-hubs.github.io/hubValidations/articles/hub-validations-class.html}{article on \verb{<hub_validations>} S3 class objects}.
}
\description{
Validate a submitted model data file submission time.
}
\examples{
hub_path <- system.file("testhubs/simple", package = "hubValidations")
file_path <- "team1-goodmodel/2022-10-08-team1-goodmodel.csv"
validate_submission_time(hub_path, file_path)
}
