% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_errors.R
\name{check_for_errors}
\alias{check_for_errors}
\title{Raise conditions stored in a \code{fhub_validations} object}
\usage{
check_for_errors(x)
}
\arguments{
\item{x}{A \code{fhub_validations} object}
}
\value{
An error if one of the elements of \code{x} is of class \code{fhub_failure} or
\code{fhub_unrecoverable_error}. \code{TRUE} invisibly otherwise.
}
\description{
This is meant to be used in CI workflows to raise conditions from
\code{fhub_validations} objects.
}
