% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target_tbl_values.R
\name{check_target_tbl_values}
\alias{check_target_tbl_values}
\title{Check that task ID columns in a target data file have valid task ID values}
\usage{
check_target_tbl_values(
  target_tbl_chr,
  target_type = c("time-series", "oracle-output"),
  file_path,
  hub_path
)
}
\arguments{
\item{target_tbl_chr}{A tibble/data.frame of the contents of the target data file
being validated. All columns should be coerced to character.}

\item{target_type}{Type of target data to retrieve matching files. One of "time-series" or
"oracle-output". Defaults to "time-series".}

\item{file_path}{A character string representing the path to the target data file
relative to the \code{target-data} directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check is only performed when the target data file contains columns that map onto
task IDs or output types defined in the hub configuration.
}
