% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_colnames.R
\name{check_tbl_colnames}
\alias{check_tbl_colnames}
\title{Check column names of model output data}
\usage{
check_tbl_colnames(tbl, round_id, file_path, hub_path = ".")
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{round_id}{character string. The round identifier.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Checks that a tibble/data.frame of data read in from the file being validated
contains the expected task ID and standard column names according the round
configuration being validated against.
}
