% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_derived_task_id_vals.R
\name{check_tbl_derived_task_id_vals}
\alias{check_tbl_derived_task_id_vals}
\title{Check derived task ID columns contain valid values}
\usage{
check_tbl_derived_task_id_vals(
  tbl,
  round_id,
  file_path,
  hub_path,
  derived_task_ids = get_hub_derived_task_ids(hub_path, round_id)
)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated. Column types must \strong{all be character}.}

\item{round_id}{character string. The round identifier.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs) to ignore. Columns for such task ids will
contain \code{NA}s. Defaults to extracting derived task IDs from \code{config_tasks}. See
\code{\link[=get_config_derived_task_ids]{get_config_derived_task_ids()}} for more details.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_failure>} condition class object.
}

If no \code{derived_task_ids} are specified, the check is skipped and a
\verb{<message/check_info>} condition class object is retuned.

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
This check is used to validate that values in any derived task ID columns
matches accepted values for each derived task ID in the config.
Given the dependence of derived task IDs on the values of other values,
it ignores the combinations of derived task ID values with those of other task IDs
and focuses only on identifying values that do not match the accepted values.
}
