% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_spl_compound_taskid_set.R
\name{check_tbl_spl_compound_taskid_set}
\alias{check_tbl_spl_compound_taskid_set}
\title{Check model output data tbl sample compound task id sets for each modeling task
match or are coarser than the expected set defined in the config.}
\usage{
check_tbl_spl_compound_taskid_set(
  tbl,
  round_id,
  file_path,
  hub_path,
  derived_task_ids = get_hub_derived_task_ids(hub_path)
)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.
Column types must \strong{all be character}.}

\item{round_id}{character string. The round identifier.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs) to ignore. Columns for such task ids will
contain \code{NA}s. Defaults to extracting derived task IDs from hub \code{task.json}. See
\code{\link[=get_hub_derived_task_ids]{get_hub_derived_task_ids()}} for more details.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
This check detects the compound task ID sets of samples, implied by the \code{output_type_id}
and task ID values, and checks them for internal consistency and compliance with
the \code{compound_taskid_set} defined for each round modeling task in the \code{tasks.json} config.
}
\details{
If the check fails, the output of the check includes an \code{errors} element,
a list of items, one for each modeling task failing validation.
The structure depends on the reason the check failed.

If the check failed because more that a single unique \code{compound_taskid_set} was found
for a given model task, the \code{errors} object will be a list with one element for each
\code{compound_taskid_set} detected and will have the following structure:
\itemize{
\item \code{tbl_comp_tids}: a compound task id set detected in the the tbl.
\item \code{output_type_ids}: The output type ID of the sample that does not contain a
single, unique value for each compound task ID.
}

If the check failed because task IDs the hub does not accept as compound task IDs
were identified as such (i.e. samples describe "finer" compound modeling tasks)
for a given model task, the \code{errors} object will be a list with the structure
described above as well as the additional following elements:
\itemize{
\item \code{config_comp_tids}: the \code{compound_taskid_set} the modeling task config expects.
\item \code{invalid_tbl_comp_tids}: the names of invalid compound task IDs.
}

The name of each element is the index identifying the config modeling task the sample is associated with \code{mt_id}.
See \href{https://docs.hubverse.io/en/latest/user-guide/sample-output-type.html}{hubverse documentation on samples}
for more details.
}
