% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_unique_round_id.R
\name{check_tbl_unique_round_id}
\alias{check_tbl_unique_round_id}
\title{Check model output data tbl contains a single unique round ID.}
\usage{
check_tbl_unique_round_id(tbl, round_id_col, file_path, hub_path)
}
\arguments{
\item{tbl}{A tibble. The contents of model output data file being validated.}

\item{round_id_col}{Character string. The name of the column containing
\code{round_id}s. Usually, the value of round property \code{round_id} in hub \code{tasks.json}
config file.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
#' Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<warning/check_failure>} condition class object.
}

If \code{round_id_from_variable: false}, check is skipped and a
\verb{<message/check_info>} condition class object is returned.
Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check model output data tbl contains a single unique round ID.
}
\details{
This check only applies to files being submitted to rounds where
\code{round_id_from_variable: true} and is skipped otherwise.
}
