% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tbl_value_col_ascending.R
\name{check_tbl_value_col_ascending}
\alias{check_tbl_value_col_ascending}
\title{Check that \code{quantile} and \code{cdf} output type values of model output data
are non-descending}
\usage{
check_tbl_value_col_ascending(tbl, file_path)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Checks that values in the \code{value} column for \code{quantile} and \code{cdf} output type
data for each unique task ID/output type combination
are non-descending when arranged by increasing \code{output_type_id} order.
Check only performed if \code{tbl} contains \code{quantile} or \code{cdf} output type data.
If not, the check is skipped and a \verb{<message/check_info>} condition class
object is returned.
}
