% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_round_id.R
\name{check_valid_round_id}
\alias{check_valid_round_id}
\title{Check whether the \code{round_id} determined for the submission is valid}
\usage{
check_valid_round_id(round_id, file_path, hub_path = ".")
}
\arguments{
\item{round_id}{character string. The round identifier.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<error/check_error>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check whether the \code{round_id} determined for the submission is valid
}
