% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Concatenate \code{hub_validations} S3 class objects}
\usage{
combine(...)
}
\arguments{
\item{...}{\code{hub_validations} S3 class objects to be concatenated.}
}
\value{
a \code{hub_validations} S3 class object.
}
\description{
Concatenate \code{hub_validations} S3 class objects
}
