% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_tbl_to_model_task.R
\name{match_tbl_to_model_task}
\alias{match_tbl_to_model_task}
\title{Match model output \code{tbl} data to their model tasks in \code{config_tasks}.}
\usage{
match_tbl_to_model_task(
  tbl,
  config_tasks,
  round_id,
  output_types = NULL,
  derived_task_ids = get_config_derived_task_ids(config_tasks, round_id),
  all_character = TRUE
)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[hubUtils:read_config]{hubUtils::read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the round's
\code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config. Ignored if hub
contains only a single round.}

\item{output_types}{Character vector of output type names to include.
Use to subset for grids for specific output types.}

\item{derived_task_ids}{Character vector of derived task ID names (task IDs whose
values depend on other task IDs) to ignore. Columns for such task ids will
contain \code{NA}s.}

\item{all_character}{Logical. Whether to return all character column.}
}
\value{
A list containing a \code{tbl_df} of model output data matched to a model
task with one element per round model task.
}
\description{
Split and match model output \code{tbl} data to their corresponding model tasks in
\code{config_tasks}. Useful for performing model task specific checks on model output.
For v3 samples, the \code{output_type_id} column is set to \code{NA} for \code{sample} outputs.
}
\examples{
hub_path <- system.file("testhubs/samples", package = "hubValidations")
tbl <- read_model_out_file(
  file_path = "flu-base/2022-10-22-flu-base.csv",
  hub_path, coerce_types = "chr"
)
config_tasks <- hubUtils::read_config(hub_path, "tasks")
match_tbl_to_model_task(tbl, config_tasks, round_id = "2022-10-22")
match_tbl_to_model_task(tbl, config_tasks,
  round_id = "2022-10-22",
  output_types = "sample"
)
}
