% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_target_validations.R
\name{new_target_validations}
\alias{new_target_validations}
\alias{as_target_validations}
\title{Create new or convert list to \code{target_validations} S3 class object}
\usage{
new_target_validations(...)

as_target_validations(x)
}
\arguments{
\item{...}{named elements to be included. Each element must be an object which
inherits from class \verb{<hub_check>}.}

\item{x}{a list of named elements. Each element must be an object which
inherits from class \verb{<hub_check>}.}
}
\value{
an S3 object of class \verb{<target_validations>}.
}
\description{
Create new or convert list to \code{target_validations} S3 class object
}
\section{Functions}{
\itemize{
\item \code{new_target_validations()}: Create new \verb{<target_validations>} S3 class object

\item \code{as_target_validations()}: Convert list to \verb{<target_validations>} S3 class object

}}
\examples{
new_target_validations()

hub_path <- system.file("testhubs/v5/target_file", package = "hubUtils")
file_path <- "time-series.csv"
new_target_validations(
  target_file_name = check_target_file_name(file_path),
  target_file_ext_valid = check_target_file_ext_valid(file_path)
)
x <- list(
  target_file_name = check_target_file_name(file_path),
  target_file_ext_valid = check_target_file_ext_valid(file_path)
)
as_target_validations(x)
file_path <- "time-series/target=wk\%20flu\%20hosp\%20rate/part-0.parquet"
new_target_validations(
  target_file_name = check_target_file_name(file_path),
  target_file_ext_valid = check_target_file_ext_valid(file_path)
)
}
