% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_check_tbl_horizon_timediff.R
\name{opt_check_tbl_horizon_timediff}
\alias{opt_check_tbl_horizon_timediff}
\title{Check time difference between values in two date columns equal a defined period.}
\usage{
opt_check_tbl_horizon_timediff(
  tbl,
  file_path,
  hub_path,
  t0_colname,
  t1_colname,
  horizon_colname = "horizon",
  timediff = lubridate::weeks()
)
}
\arguments{
\item{tbl}{a tibble/data.frame of the contents of the file being validated.}

\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{t0_colname}{Character string. The name of the time zero date column.}

\item{t1_colname}{Character string. The name of the time zero + 1 time step date column.}

\item{horizon_colname}{Character string. The name of the horizon column.
Defaults to \code{"horizon"}.}

\item{timediff}{an object of class \code{lubridate} \code{\linkS4class{Period}} and length 1.
The period of a single horizon. Default to 1 week.}
}
\value{
Depending on whether validation has succeeded, one of:
\itemize{
\item \verb{<message/check_success>} condition class object.
\item \verb{<warning/check_failure>} condition class object.
}

Returned object also inherits from subclass \verb{<hub_check>}.
}
\description{
Check time difference between values in two date columns equal a defined period.
}
