% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hub_validations}
\alias{print.hub_validations}
\title{Print results of \code{validate_...()} function as a bullet list}
\usage{
\method{print}{hub_validations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hub_validations} or any of it's subclasses.}

\item{...}{Unused argument present for class consistency}
}
\description{
Print results of \code{validate_...()} function as a bullet list
}
