% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_out_file.R
\name{read_model_out_file}
\alias{read_model_out_file}
\title{Read a model output file}
\usage{
read_model_out_file(
  file_path,
  hub_path = ".",
  coerce_types = c("hub", "chr", "none"),
  output_type_id_datatype = c("from_config", "auto", "character", "double", "integer",
    "logical", "Date")
)
}
\arguments{
\item{file_path}{character string. Path to the file being validated relative to
the hub's model-output directory.}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{s3_bucket()}}
or \code{\link[hubData:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{coerce_types}{character. What to coerce column types to on read.
\itemize{
\item \code{hub}: (default) read in (\code{csv}) or coerce (\code{parquet}, \code{arrow}) to hub
schema.
When coercing data types using the \code{hub} schema, the \code{output_type_id_datatype}
can also be used to set the \code{output_type_id} column data type manually.
\item \code{chr}: read in (\code{csv}) or coerce (\code{parquet}, \code{arrow}) all columns to character.
\item \code{none}: No coercion. Use \code{arrow} \verb{read_*} function defaults.
}}

\item{output_type_id_datatype}{character string. One of \code{"from_config"}, \code{"auto"},
\code{"character"}, \code{"double"}, \code{"integer"}, \code{"logical"}, \code{"Date"}.
Defaults to \code{"from_config"} which uses the setting in the \code{output_type_id_datatype}
property in the \code{tasks.json} config file if available. If the property is
not set in the config, the argument falls back to \code{"auto"} which determines
the  \code{output_type_id} data type automatically from the \code{tasks.json}
config file as the simplest data type required to represent all output
type ID values across all output types in the hub.
Other data type values can be used to override automatic determination.
Note that attempting to coerce \code{output_type_id} to a data type that is
not valid for the data (e.g. trying to coerce\code{"character"} values to
\code{"double"}) will likely result in an error or potentially unexpected
behaviour so use with care.}
}
\value{
a tibble of contents of the model output file.
}
\description{
Read a model output file
}
