% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_check.R
\name{try_check}
\alias{try_check}
\title{Wrap check expression to capture check execution errors}
\usage{
try_check(expr)
}
\arguments{
\item{expr}{check function expression to run.}
}
\value{
If \code{expr} executes correctly, the output of \code{expr} is returned. If
The execution error message is attached as attribute \code{msg}.
}
\description{
Wrap check expression to capture check execution errors
}
