% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_data.R
\name{validate_model_data}
\alias{validate_model_data}
\title{Validate model data file}
\usage{
validate_model_data(data_file, data_schema)
}
\arguments{
\item{data_file}{Path to the \code{.csv} file}

\item{data_schema}{Path to the \code{.yml} schema file}
}
\value{
An object of class \code{fhub_validations}.
}
\description{
Validate model data file
}
\examples{
validate_model_data(
  system.file(
    "testdata", "example-model", "2021-07-26-example-model.csv",
    package = "ForecastHubValidations"
  ),
  system.file(
    "testdata", "schema-forecast.yml",
    package = "ForecastHubValidations"
  )
)
}
