% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pr.R
\name{validate_pr}
\alias{validate_pr}
\title{Validate forecast hub from pull request}
\usage{
validate_pr(
  gh_repo,
  pr_number,
  data_folder,
  metadata_folder,
  local = identical(Sys.getenv("GITHUB_ACTIONS"), "true") &&
    identical(Sys.getenv("GITHUB_REPOSITORY"), gh_repo),
  ...
)
}
\arguments{
\item{gh_repo}{GitHub repository address in the format \code{username/repo}}

\item{pr_number}{Number of the pull request to validate}

\item{data_folder}{The path to the folder containing forecasts}

\item{metadata_folder}{The path to the folder containing metadata}

\item{local}{Logical. Is this function called from your local computer or
from a continuous integration system. By default, it tries to guess the
answer based on the values of some environment variables}

\item{...}{Arguments passed to \code{\link[=validate_repository]{validate_repository()}}}
}
\value{
An object of class \code{fhub_validations}.
}
\description{
Validate forecast hub from pull request
}
\examples{
\dontrun{
validate_pr(
  "epiforecasts/covid19-forecast-hub-europe",
  536,
  data_folder = "data-processed"
)
}
}
