% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pr.R
\name{validate_pr}
\alias{validate_pr}
\title{Validate pull request submission}
\usage{
validate_pr(hub_path = ".", gh_repo, pr_number)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{gh_repo}{GitHub repository address in the format \code{username/repo}}

\item{pr_number}{Number of the pull request to validate}
}
\value{
An object of class \code{hub_validations}.
}
\description{
Validate pull request submission
}
\examples{
\dontrun{
validate_pr(
  hub_path = "."
  gh_repo = "Infectious-Disease-Modeling-Hubs/ci-testhub-simple",
  pr_number = 3
)
}
}
