% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pr.R
\name{validate_pr}
\alias{validate_pr}
\title{Validate Pull Request}
\usage{
validate_pr(
  hub_path = ".",
  gh_repo,
  pr_number,
  round_id_col = NULL,
  validations_cfg_path = NULL,
  skip_submit_window_check = FALSE
)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubUtils:s3_bucket]{s3_bucket()}}
or \code{\link[hubUtils:gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{gh_repo}{GitHub repository address in the format \code{username/repo}}

\item{pr_number}{Number of the pull request to validate}

\item{round_id_col}{Character string. The name of the column containing
\code{round_id}s. Only required if files contain a column that contains \code{round_id}
details but has not been configured via \code{round_id_from_variable: true} and
\verb{round_id:} in in hub \code{tasks.json} config file.}

\item{validations_cfg_path}{Path to \code{validations.yml} file. If \code{NULL}
defaults to \code{hub-config/validations.yml}.}

\item{skip_submit_window_check}{Logical. Whether to skip the submission window check.}
}
\value{
An object of class \code{hub_validations}.
}
\description{
Validate Pull Request
}
\examples{
\dontrun{
validate_pr(
  hub_path = "."
  gh_repo = "Infectious-Disease-Modeling-Hubs/ci-testhub-simple",
  pr_number = 3
)
}
}
